# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Remitly 2025 internship task",
        "title": "Remitly SWIFT API 2025",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/swift-codes": {
            "post": {
                "description": "Creates a bank",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Creates a bank",
                "parameters": [
                    {
                        "description": "Bank payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.BankPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/swift-codes/country/{countryISO2code}": {
            "get": {
                "description": "Gets all banks with given Country ISO2 Code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Gets all banks with given Country ISO2 Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country ISO2 Code",
                        "name": "countryISO2code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.AllBanks"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/swift-codes/{swift-code}": {
            "get": {
                "description": "Gets a bank by SWIFT code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Gets a bank by SWIFT code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SWIFT Code",
                        "name": "swift-code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns either a BankHeadquarter or BankBranch. See the API documentation for details.",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a bank by SWIFT code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banks"
                ],
                "summary": "Deletes a bank by SWIFT code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SWIFT Code",
                        "name": "swift-code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requests.BankPayload": {
            "type": "object",
            "required": [
                "bankName",
                "countryISO2",
                "countryName",
                "isHeadquarter",
                "swiftCode"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255
                },
                "bankName": {
                    "type": "string",
                    "maxLength": 255
                },
                "countryISO2": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string",
                    "maxLength": 255
                },
                "isHeadquarter": {
                    "type": "boolean"
                },
                "swiftCode": {
                    "type": "string"
                }
            }
        },
        "responses.AllBanks": {
            "type": "object",
            "properties": {
                "countryISO2": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "swiftCodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.BankShort"
                    }
                }
            }
        },
        "responses.BankShort": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "countryISO2": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "isHeadquarter": {
                    "type": "boolean"
                },
                "swiftCode": {
                    "type": "string"
                }
            }
        },
        "responses.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "responses.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}